+++ src/config/mod.rs
@@ -94,6 +94,7 @@ pub fn migrate_legacy_config(
         let profile = structs::Profile {
             filters: legacy::migrate_filters(filters)?,
             imports: Vec::new(),
+            overrides: None,
             mods: {
                 let mut mods = HashMap::new();
                 for mod_ in legacy_mods {
@@ -129,6 +130,7 @@ pub fn migrate_legacy_config(
         let profile = structs::Profile {
             filters: structs::Filters::empty(),
             imports: Vec::new(),
+            overrides: None,
             mods: HashMap::new(),
             shaders: HashMap::new(),
             modpacks: HashMap::from([(legacy_modpack.name.clone(), source)]),